jacobi example converted to use op2 mpi

+++ jac1/jac_kernels.cpp
@@ -12,7 +12,7 @@
                             
 // global constants         
                             
-extern float alpha;         
+extern double alpha;         
                             
 // user kernel files
